marketing/requirements.txt
google-adk
google-genai
python-dotenv
beautifulsoup4
requests
pandas
plotly

